+++ src/main/java/ch/newinstance/plugin/mavendependencychecker/parser/DependencyParser.java
@@ -1,8 +1,5 @@
 package ch.newinstance.plugin.mavendependencychecker.parser;
 
-import com.intellij.openapi.module.Module;
-import com.intellij.openapi.roots.ModuleRootManager;
-import com.intellij.openapi.roots.ProjectRootManager;
 import com.intellij.psi.PsiFile;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.model.Dependency;
@@ -10,6 +7,7 @@ import org.apache.maven.model.Model;
 import org.apache.maven.model.Plugin;
 import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
 import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
+import org.jetbrains.idea.maven.model.MavenArtifact;
 import org.jetbrains.idea.maven.model.MavenPlugin;
 import org.jetbrains.idea.maven.project.MavenProject;
 import org.jetbrains.idea.maven.project.MavenProjectsManager;
@@ -18,9 +16,7 @@ import java.io.IOException;
 import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Properties;
 
 public class DependencyParser {
@@ -52,22 +48,9 @@ public class DependencyParser {
         }
     }
 
-    public Map<String, String> parseModuleDependencies() {
-        Module module = ProjectRootManager.getInstance(pomFile.getProject()).getFileIndex().getModuleForFile(pomFile.getVirtualFile());
-        if (module == null) {
-            return Collections.emptyMap();
-        }
-
-        Map<String, String> libraryMap = new HashMap<>();
-        ModuleRootManager.getInstance(module).orderEntries().forEachLibrary(library -> {
-            String[] libraryParts = StringUtils.split(library.getName(), ":");
-            if (libraryParts == null || libraryParts.length < 4) {
-                return false;
-            }
-            libraryMap.put(libraryParts[1].trim() + ":" + libraryParts[2].trim(), libraryParts[3].trim());
-            return true;
-        });
-        return libraryMap;
+    public List<MavenArtifact> parseModuleDependencies() {
+        List<MavenProject> mavenProjects = MavenProjectsManager.getInstance(pomFile.getProject()).getProjects();
+        return mavenProjects.get(0).getDependencies();
     }
 
     public List<Plugin> parseMavenPlugins() {